requirements.txt
# Core web framework
Flask>=2.3.2

# Image processing
Pillow>=9.5.0

# PyTorch and TorchVision (for models and transforms)
torch>=2.0.0
torchvision>=0.15.0

# For downloading ImageNet labels
requests>=2.31.0
